requirements.txt
aiofiles==22.1.0
aiogram==3.0.0b6
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
APScheduler==3.6.3
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==3.1.0
cssselect==1.2.0
dataclasses-json==0.5.7
fake-useragent==1.1.3
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
langchain==0.0.151
lxml==4.9.2
magic-filter==1.0.9
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.5
openapi-schema-pydantic==1.2.4
packaging==23.1
parse==1.19.0
pexpect==4.8.0
ptyprocess==0.7.0
pydantic==1.10.7
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
pyTelegramBotAPI==4.11.0
python-telegram-bot==13.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.29.0
requests-html==0.10.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.11
tenacity==8.2.2
tornado==6.3.1
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
w3lib==2.1.1
websockets==10.4
yarl==1.9.2
zipp==3.15.0
